backend/requirements.txt
# Save-It.AI Backend Dependencies
# Install with: pip install -r requirements.txt

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
starlette>=0.27.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# Authentication & Security
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Environment & Config
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# AI & LLM
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.0.0

# IoT & Device Communication
aiomqtt>=2.0.0
pymodbus>=3.6.0
paho-mqtt>=2.0.0

# Caching & Background Tasks
redis>=5.0.0
apscheduler>=3.10.0

# HTTP & Async
httpx>=0.25.0
aiofiles>=23.0.0

# File uploads
python-multipart>=0.0.6

# Email
emails>=0.6
jinja2>=3.1.0

# Date utilities
python-dateutil>=2.8.0

# Financial calculations
numpy-financial>=1.0.0

# Excel/spreadsheet support
openpyxl>=3.1.0

# PDF generation
reportlab>=4.0.0

# System monitoring
psutil>=5.9.0

# Async MQTT
paho-mqtt>=2.0.0
